// src/generated/t_tech/invest/grpc/instruments.ts
export interface TradingSchedule {
  /** Наименование торговой площадки. */
  exchange: string;
  /** Массив с торговыми и неторговыми днями. */
  days: TradingDay[];
}

/** Информация о времени торгов. */
export interface TradingDay {
  /** Дата. */
  date:
    | Date
    | undefined;
  /** Признак торгового дня на бирже. */
  isTradingDay: boolean;
  /** Время начала торгов по UTC. */
  startTime:
    | Date
    | undefined;
  /** Время окончания торгов по UTC. */
  endTime:
    | Date
    | undefined;
  /** Время начала аукциона открытия по UTC. */
  openingAuctionStartTime:
    | Date
    | undefined;
  /** Время окончания аукциона закрытия по UTC. */
  closingAuctionEndTime:
    | Date
    | undefined;
  /** Время начала аукциона открытия вечерней сессии по UTC. */
  eveningOpeningAuctionStartTime:
    | Date
    | undefined;
  /** Время начала вечерней сессии по UTC. */
  eveningStartTime:
    | Date
    | undefined;
  /** Время окончания вечерней сессии по UTC. */
  eveningEndTime:
    | Date
    | undefined;
  /** Время начала основного клиринга по UTC. */
  clearingStartTime:
    | Date
    | undefined;
  /** Время окончания основного клиринга по UTC. */
  clearingEndTime:
    | Date
    | undefined;
  /** Время начала премаркета по UTC. */
  premarketStartTime:
    | Date
    | undefined;
  /** Время окончания премаркета по UTC. */
  premarketEndTime:
    | Date
    | undefined;
  /** Время начала аукциона закрытия по UTC. */
  closingAuctionStartTime:
    | Date
    | undefined;
  /** Время окончания аукциона открытия по UTC. */
  openingAuctionEndTime:
    | Date
    | undefined;
  /** Торговые интервалы. */
  intervals: TradingInterval[];
}
// ...
export interface TradingInterval {
  /** Название интервала. */
  type: string;
  /** Интервал. */
  interval: TradingInterval_TimeInterval | undefined;
}

export interface TradingInterval_TimeInterval {
  /** Время начала интервала. */
  startTs:
    | Date
    | undefined;
  /** Время окончания интервала. */
  endTs: Date | undefined;
}
// ...
export const TradingSchedule: MessageFns<TradingSchedule> = {
  encode(message: TradingSchedule, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.exchange !== "") {
      writer.uint32(10).string(message.exchange);
    }
    for (const v of message.days) {
      TradingDay.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TradingSchedule {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTradingSchedule();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.exchange = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.days.push(TradingDay.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TradingSchedule {
    return {
      exchange: isSet(object.exchange) ? globalThis.String(object.exchange) : "",
      days: globalThis.Array.isArray(object?.days) ? object.days.map((e: any) => TradingDay.fromJSON(e)) : [],
    };
  },

  toJSON(message: TradingSchedule): unknown {
    const obj: any = {};
    if (message.exchange !== "") {
      obj.exchange = message.exchange;
    }
    if (message.days?.length) {
      obj.days = message.days.map((e) => TradingDay.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<TradingSchedule>): TradingSchedule {
    return TradingSchedule.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TradingSchedule>): TradingSchedule {
    const message = createBaseTradingSchedule();
    message.exchange = object.exchange ?? "";
    message.days = object.days?.map((e) => TradingDay.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const TradingInterval: MessageFns<TradingInterval> = {
  encode(message: TradingInterval, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== "") {
      writer.uint32(10).string(message.type);
    }
    if (message.interval !== undefined) {
      TradingInterval_TimeInterval.encode(message.interval, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TradingInterval {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTradingInterval();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.type = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.interval = TradingInterval_TimeInterval.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TradingInterval {
    return {
      type: isSet(object.type) ? globalThis.String(object.type) : "",
      interval: isSet(object.interval) ? TradingInterval_TimeInterval.fromJSON(object.interval) : undefined,
    };
  },

  toJSON(message: TradingInterval): unknown {
    const obj: any = {};
    if (message.type !== "") {
      obj.type = message.type;
    }
    if (message.interval !== undefined) {
      obj.interval = TradingInterval_TimeInterval.toJSON(message.interval);
    }
    return obj;
  },

  create(base?: DeepPartial<TradingInterval>): TradingInterval {
    return TradingInterval.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TradingInterval>): TradingInterval {
    const message = createBaseTradingInterval();
    message.type = object.type ?? "";
    message.interval = (object.interval !== undefined && object.interval !== null)
      ? TradingInterval_TimeInterval.fromPartial(object.interval)
      : undefined;
    return message;
  },
};

function createBaseTradingInterval_TimeInterval(): TradingInterval_TimeInterval {
  return { startTs: undefined, endTs: undefined };
}

export const TradingInterval_TimeInterval: MessageFns<TradingInterval_TimeInterval> = {
  encode(message: TradingInterval_TimeInterval, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.startTs !== undefined) {
      Timestamp.encode(toTimestamp(message.startTs), writer.uint32(10).fork()).join();
    }
    if (message.endTs !== undefined) {
      Timestamp.encode(toTimestamp(message.endTs), writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TradingInterval_TimeInterval {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTradingInterval_TimeInterval();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.startTs = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.endTs = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TradingInterval_TimeInterval {
    return {
      startTs: isSet(object.startTs)
        ? fromJsonTimestamp(object.startTs)
        : isSet(object.start_ts)
        ? fromJsonTimestamp(object.start_ts)
        : undefined,
      endTs: isSet(object.endTs)
        ? fromJsonTimestamp(object.endTs)
        : isSet(object.end_ts)
        ? fromJsonTimestamp(object.end_ts)
        : undefined,
    };
  },

  toJSON(message: TradingInterval_TimeInterval): unknown {
    const obj: any = {};
    if (message.startTs !== undefined) {
      obj.startTs = message.startTs.toISOString();
    }
    if (message.endTs !== undefined) {
      obj.endTs = message.endTs.toISOString();
    }
    return obj;
  },

  create(base?: DeepPartial<TradingInterval_TimeInterval>): TradingInterval_TimeInterval {
    return TradingInterval_TimeInterval.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TradingInterval_TimeInterval>): TradingInterval_TimeInterval {
    const message = createBaseTradingInterval_TimeInterval();
    message.startTs = object.startTs ?? undefined;
    message.endTs = object.endTs ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}